requirements.txt
pretty_midi==0.2.8
tensorflow_gpu==2.1.0
MIDIUtil==1.2.1
numpy==1.18.1
matplotlib==3.1.3
Keras==2.3.1
pandas==1.0.1
tqdm==4.42.1
magenta==1.2.2
algo_gen==0.16
music21==5.7.2
scikit_learn==0.22.1
